+++ src/Components/Containers/Grid/index.tsx
@@ -3,7 +3,7 @@ import UpRow from './UpRow';
 import UpCol from './UpCol';
 
 export {
-  UpGrid,
-  UpRow,
-  UpCol
+    UpGrid,
+    UpRow,
+    UpCol
 };