+++ src/Components/Containers/NavTab/index.stories.tsx
@@ -15,15 +15,35 @@ export const GeneralUse =
         <UpNavTab tabs={
             [
                 {
-                    content: <UpPanel type="info" disableAutoIntentIcon={false} message="Information sur ..." />,
-                    head: "Tab 1"
+                    content: <UpPanel type='info' disableAutoIntentIcon={false} message='Information sur ...' />,
+                    head: 'Tab 1'
                 },
                 {
-                    content: <UpPanel type="warning" disableAutoIntentIcon={false} message="Attention sur ..." />,
-                    head: "Tab 2"
+                    content: <UpPanel type='warning' disableAutoIntentIcon={false} message='Attention sur ...' />,
+                    head: 'Tab 2'
                 }, {
-                    content: <UpPanel type="danger" disableAutoIntentIcon={false} message="Erreur sur ..." />,
-                    head: "Tab 3"
+                    content: <UpPanel type='danger' disableAutoIntentIcon={false} message='Erreur sur ...' />,
+                    head: 'Tab 3'
+                }
+            ]
+        }
+        />
+    );
+
+export const SelectedTabOnLoad =
+    () => (
+        <UpNavTab tabs={
+            [
+                {
+                    content: <UpPanel type='info' disableAutoIntentIcon={false} message='Information sur ...' />,
+                    head: 'Tab 1'
+                },
+                {
+                    content: <UpPanel type='warning' disableAutoIntentIcon={false} message='Attention sur ...' />,
+                    head: 'Tab 2'
+                }, {
+                    content: <UpPanel type='danger' disableAutoIntentIcon={false} message='Erreur sur ...' />,
+                    head: 'Tab 3'
                 }
             ]
         }
@@ -31,20 +51,21 @@ export const GeneralUse =
         />
     );
 
+
 export const OnShowLoadType =
     () => (
         <UpNavTab tabs={
             [
                 {
-                    content: <UpPanel type="info" disableAutoIntentIcon={false} message="Information sur ..." />,
-                    head: "Tab 1"
+                    content: <UpPanel type='info' disableAutoIntentIcon={false} message='Information sur ...' />,
+                    head: 'Tab 1'
                 },
                 {
-                    content: <UpPanel type="warning" disableAutoIntentIcon={false} message="Attention sur ..." />,
-                    head: "Tab 2"
+                    content: <UpPanel type='warning' disableAutoIntentIcon={false} message='Attention sur ...' />,
+                    head: 'Tab 2'
                 }, {
-                    content: <UpPanel type="danger" disableAutoIntentIcon={false} message="Erreur sur ..." />,
-                    head: "Tab 3"
+                    content: <UpPanel type='danger' disableAutoIntentIcon={false} message='Erreur sur ...' />,
+                    head: 'Tab 3'
                 }
             ]
         }
@@ -52,23 +73,25 @@ export const OnShowLoadType =
         />
     );
 
-export const PassCallBackToUpNavTab =
+export const CallBackUpNavTab =
     () => (
         <UpNavTab tabs={
             [
                 {
-                    content: <UpPanel type="info" disableAutoIntentIcon={false} message="Information sur ..." />,
-                    head: "Tab 1"
+                    content: <UpPanel type='info' disableAutoIntentIcon={false} message='Information sur ...' />,
+                    head: 'Tab 1'
                 },
                 {
-                    content: <UpPanel type="warning" disableAutoIntentIcon={false} message="Attention sur ..." />,
-                    head: "Tab 2"
+                    content: <UpPanel type='warning' disableAutoIntentIcon={false} message='Attention sur ...' />,
+                    head: 'Tab 2'
                 }, {
-                    content: <UpPanel type="danger" disableAutoIntentIcon={false} message="Erreur sur ..." />,
-                    head: "Tab 3"
+                    content: <UpPanel type='danger' disableAutoIntentIcon={false} message='Erreur sur ...' />,
+                    head: 'Tab 3'
                 }
             ]
         }
             onSelectedTabChanged={(selectTabKey, tab) => console.log(`You are in tab ${selectTabKey}`)}
         />
     );
+
+CallBackUpNavTab.storyName = 'Callback To Execute On Tab Changing';